+++ persistence-modules/read-only-transactions/src/main/java/com/baeldung/readonlytransactions/mysql/entities/Transaction.java
@@ -1,4 +1,4 @@
-package com.baeldung.read_only_transactions.h2;
+package com.baeldung.readonlytransactions.mysql.entities;
 
 import javax.persistence.*;
