+++ persistence-modules/read-only-transactions/src/main/java/com/baeldung/readonlytransactions/mysql/spring/ReadOnlyContext.java
@@ -1,4 +1,4 @@
-package com.baeldung.read_only_transactions.mysql.spring;
+package com.baeldung.readonlytransactions.mysql.spring;
 
 import java.util.concurrent.atomic.AtomicInteger;
